# openapi.yaml
openapi: 3.0.3
info:
  title: 辰星认证中枢 API
  version: 1.0.0
  description: |
    辰星通行证用户认证、OAuth 2.0/OIDC 和统一用户管理 API。
    当前定义以仓库实际路由为准。运行期 Issuer 持久化在 PostgreSQL；未配置时仅健康检查、静态前端和初始化状态可用，其他协议路由由运行时门禁拒绝。

    请求超时（`REQUEST_TIMEOUT_SECONDS`）按路径分流错误信封（Issue #423）：
    `/oauth/*` 协议端点返回 RFC 6749 `503 temporarily_unavailable`（`OAuthErrorResponse`），
    不得改写为内部 API 信封；其余应用路由返回 `504` + `ErrorResponse`（`code=request_timeout`）。
    健康检查与静态 SPA fallback 不套用该请求超时。
servers:
  - url: http://127.0.0.1:3000
    description: 本地开发
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [User Auth/Account]
      summary: 用户登录
      description: |
        错误码：
        - 401 invalid_credentials / invalid_factor：口令或 TOTP 验证码错误，或已被限流。
        - 503 factor_key_unavailable：请求带 `totp_code` 且该账号的 TOTP 种子由已退役的
          加密 kid 保护，服务端无法解密。重试无效，需要管理员重置因子。该结果不消耗
          认证失败额度。
      operationId: loginUser
      parameters:
        - { $ref: '#/components/parameters/SessionResponseMode' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: 登录成功，同时设置 Session 和 CSRF Cookie
          headers:
            Set-Cookie: { $ref: '#/components/headers/SetCookie' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '202':
          description: 密码正确但需要首次绑定或完成认证因子。通过 HttpOnly login ticket 与 holder Cookie 继续流程，不在 JSON 响应中返回 ticket
          headers:
            Set-Cookie: { $ref: '#/components/headers/SetCookie' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PendingLoginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /api/v1/auth/authorized-apps:
    get:
      tags: [User Auth/Authorized Apps]
      summary: 查询当前用户已授权的 OAuth 应用
      operationId: listAuthorizedApps
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: 当前用户授权记录的非敏感摘要，不包含 Secret、哈希、重定向地址或 Token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthorizedOAuthAppList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/authorized-apps/{client_id}:
    delete:
      tags: [User Auth/Authorized Apps]
      summary: 撤销当前用户对 OAuth 应用的授权
      operationId: revokeAuthorizedApp
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: client_id, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/CsrfHeader' }
      responses:
        '204': { description: 撤销成功或授权不存在，操作幂等 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /oauth/token:
    post:
      tags: [OAuth/OIDC]
      summary: 交换授权码或刷新 Token
      operationId: token
      description: |
        表单编码。Client 认证为 HTTP Basic 或 body 二选一。
        请求超时返回 503 `temporarily_unavailable`（RFC 6749 信封），不会返回内部 API 的 504 `request_timeout`（Issue #423）。
      security: [{ basicClientAuth: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token 响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400': { $ref: '#/components/responses/OAuthBadRequest' }
        '401': { $ref: '#/components/responses/OAuthUnauthorized' }
        '429': { $ref: '#/components/responses/OAuthTooManyRequests' }
        '503': { $ref: '#/components/responses/OAuthServiceUnavailable' }
...
  /api/v1/admin/bootstrap:
    post:
      tags: [Admin Auth]
      summary: 初始化首个 Owner 用户
      description: Issuer 已配置但 Owner 尚未初始化时允许匿名调用；按可信源 IP 限制为每分钟 5 次。缺少可信源地址时遵循服务端 missing_source_ip_policy，限流依赖不可用时 fail closed。
      operationId: bootstrapAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BootstrapOwnerInput' }
      responses:
        '201':
          description: 初始化成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BootstrapOwnerResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/admin/bootstrap/status:
    get:
      tags: [Admin Auth]
      summary: 查询未初始化状态
      description: Issuer 未配置时返回 503 issuer_not_configured；Issuer 已配置且 Owner 未初始化时返回 initialized=false；实例已有 Owner 后返回与未知路径一致的 404。
      operationId: adminBootstrapStatus
      responses:
        '200':
          description: 初始化状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BootstrapStatusResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /api/v1/admin/settings/issuer:
    get:
      tags: [Admin]
      summary: 查询 OIDC Issuer 状态
      description: Owner 专属。返回持久化记录、当前进程加载快照和运行阶段，不返回秘密材料。
      operationId: getIssuerSetting
      security:
        - adminBearer: []
        - sessionCookie: []
      responses:
        '200':
          description: Issuer 的持久化与运行时状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssuerSettingResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [Admin]
      summary: 创建或更新 OIDC Issuer
      description: |
        Owner 专属，浏览器会话需要 CSRF。value 必须是无 userinfo、路径、查询和片段的
        http(s) 根 URL。expected_generation 用于 CAS 并发保护；修改已有值时 confirm 必须为 true。
        写入与 issuer_configure/issuer_update 审计在同一事务中提交，当前进程立即热生效，其他副本按 generation 轮询收敛。
      operationId: updateIssuerSetting
      security:
        - sessionCookie: []
        - adminBearer: []
      parameters:
        - { $ref: '#/components/parameters/CsrfHeader' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateIssuerSetting' }
      responses:
        '200':
          description: Issuer 已持久化并加载到当前进程
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssuerSettingResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    adminBearer: { type: http, scheme: bearer, description: 系统 ADMIN_TOKEN 自动化通道，权限等价于 Owner 且无用户 ID，不需要浏览器 CSRF。ADMIN_TOKEN 为空时整个管理面关闭：Bearer 与 sessionCookie 两条通道都被拒绝，已初始化的管理接口统一返回 403 admin_disabled；唯一例外是不存在 Owner 时公开的首个 Owner 初始化接口 }
    basicClientAuth: { type: http, scheme: basic }
    sessionCookie: { type: apiKey, in: cookie, name: __Host-chenxing_session, description: HTTPS 浏览器部署使用的 host-only Session Cookie；仅 loopback HTTP 本地开发使用不带前缀的兼容名称 }
    chenxingSession: { type: apiKey, in: header, name: X-Chenxing-Session, description: 仅开发兼容配置 OAUTH_SESSION_HEADER_ENABLED=true 时可用；不能替代 pending、Session 绑定、CSRF 或 consent }
  parameters:
...
    SessionResponseMode: { name: X-Chenxing-Session-Mode, in: header, required: false, description: 非浏览器兼容模式。仅在 SESSION_TOKEN_RESPONSE_ENABLED=true 且值为 token 时将 session_id 放入 JSON；浏览器不得发送此请求头。, schema: { type: string, enum: [token] } }
    CsrfHeader: { name: X-CSRF-Token, in: header, required: true, description: 浏览器 Session 写操作必需；ADMIN_TOKEN Bearer 请求不需要。ADMIN_TOKEN 为空时整个管理面关闭，管理请求一律 403 admin_disabled，本请求头不再生效, schema: { type: string } }
...
  headers:
    SetCookie: { description: Session、CSRF 或 pending-login Cookie, schema: { type: string } }
  responses:
    BadRequest: { description: 请求错误, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Unauthorized: { description: 未认证或凭据无效, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Forbidden: { description: 无权限（admin_forbidden）；ADMIN_TOKEN 未配置时管理接口统一返回 admin_disabled, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Conflict: { description: 资源冲突（包括套餐 code 冲突、已归档套餐不能设为默认和 Passkey 禁用保护）, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    NotFound: { description: 资源不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    TooManyRequests: { description: 认证、OAuth 配额或 QPS 限流触发, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    ServiceUnavailable: { description: 依赖或必要能力暂不可用, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    InternalServerError: { description: 服务端内部故障, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    OAuthBadRequest: { description: OAuth 请求错误，响应遵守 RFC 6749/7009, content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } } }
    OAuthUnauthorized:
      description: OAuth 客户端或 Bearer 凭据无效
      headers:
        WWW-Authenticate: { description: RFC 6749/6750 challenge, schema: { type: string } }
      content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } }
    OAuthTooManyRequests: { description: OAuth 限流或容量保护触发, content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } } }
    OAuthServiceUnavailable:
      description: |
        OAuth 依赖暂时不可用，或 `/oauth/*` 协议端点触发请求超时（Issue #423）。
        响应始终为 RFC 6749 信封（`error=temporarily_unavailable`），不是内部 `{code, message}`。
      content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } }
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    OAuthErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string, description: RFC 6749/6750/7009 error code }
        error_description: { type: string, description: Human-readable non-sensitive description }
...
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier: { type: string, maxLength: 254, description: 用户名或邮箱；超过 254 字符按凭据错误拒绝 }
        password: { type: string, format: password, minLength: 10, maxLength: 128, description: 超过 128 字符按凭据错误拒绝，不返回独立错误码 }
        totp_code: { type: string, pattern: '^[0-9]{6}$' }
    PendingLoginResponse:
      type: object
      required: [status, methods]
      properties:
        status: { type: string, enum: [factor_setup_required, factor_required] }
        methods: { type: array, items: { type: string, enum: [totp, passkey] } }
...
    LoginResponse:
      type: object
      required: [expires_at]
      description: 浏览器安全的登录响应。Session token 默认不出现在 JSON 中，客户端应使用 HttpOnly Cookie 并通过 auth/status 和 auth/me 确认登录状态。
      properties:
        session_id: { type: string, description: '仅在 SESSION_TOKEN_RESPONSE_ENABLED=true 且请求显式发送 X-Chenxing-Session-Mode: token 时返回；浏览器响应默认不包含此字段' }
        expires_at: { type: string, format: date-time }
...
    AuthorizedOAuthApp:
      type: object
      required: [client_id, client_name, scopes, updated_at]
      properties:
        client_id: { type: string }
        client_name: { type: string, description: OAuth 应用显示名称 }
        scopes: { type: array, items: { type: string } }
        updated_at: { type: string, format: date-time }
    AuthorizedOAuthAppList:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AuthorizedOAuthApp' } }
...
    TokenRequest:
      type: object
      required: [grant_type]
      properties:
        grant_type: { type: string, enum: [authorization_code, refresh_token] }
        code: { type: string }
        redirect_uri: { type: string, format: uri }
        code_verifier: { type: string }
        refresh_token: { type: string }
        scope:
          type: string
          description: refresh_token grant 可省略；空白字符串按省略处理并保留原授权 scope，非空值只能缩小原授权范围。
        client_id: { type: string }
        client_secret: { type: string, format: password }
    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, format: int64 }
        scope: { type: string }
        refresh_token: { type: string }
        id_token: { type: string, description: "JWT ID Token（scope 含 openid 时返回）" }
...
    BootstrapOwnerInput:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 64, pattern: '^[A-Za-z0-9._-]+$' }
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 10 }
    BootstrapOwnerResponse:
      type: object
      required: [id, username, email, role]
      properties:
        id: { type: integer, format: int64, minimum: 1 }
        username: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [owner] }
    BootstrapStatusResponse:
      type: object
      required: [initialized]
      properties:
        initialized: { type: boolean }
    IssuerRecordResponse:
      type: object
      required: [value, generation, updated_at]
      properties:
        value: { type: string, format: uri, description: 无尾斜杠的规范化 OIDC 根 URL }
        generation: { type: integer, format: int64, minimum: 1 }
        updated_at: { type: string, format: date-time }
    IssuerSettingResponse:
      type: object
      required: [persisted, loaded, phase]
      properties:
        persisted:
          allOf: [{ $ref: '#/components/schemas/IssuerRecordResponse' }]
          nullable: true
        loaded:
          allOf: [{ $ref: '#/components/schemas/IssuerRecordResponse' }]
          nullable: true
        phase: { type: string, enum: [awaiting_issuer, issuer_loaded, issuer_invalid] }
    UpdateIssuerSetting:
      type: object
      required: [value, expected_generation, confirm]
      properties:
        value: { type: string, format: uri, maxLength: 2048 }
        expected_generation: { type: integer, format: int64, minimum: 0 }
        confirm: { type: boolean, description: 已有 Issuer 改值时必须为 true }
...